Extract PrefixedTableNames
